split util into s3_utils and file_utils, updated imports, comment changes
