+++ datalab/datalab_session/analysis/get_tif.py
@@ -1,4 +1,5 @@
-from datalab.datalab_session.util import create_tif, get_fits, add_file_to_bucket, key_exists, get_s3_url
+from datalab.datalab_session.file_utils import create_tif, get_fits
+from datalab.datalab_session.s3_utils import key_exists, add_file_to_bucket, get_s3_url
 
 def get_tif(input: dict):
   """
